Update subscribe message with the appropriate remote-addr
On a newly provisioned system, subscribers can be up before DiscoveryServer or haproxy, in such
a case socket connect fails and hence local ip address defaults to localhost (127.0.0.1), retry
socket connect for every subscribe when ip is still set to localhost.

This fix is a must when policies are configured to filter or group subscribers and publishers.

This fix is for c++ client only, following commit for python clients.

Change-Id: I5ae8c3a3bf362c3a99f29d04cac2b0cfc137ad4c
Partial-Bug:1484861