Removed Redis-Query connection from process_status for analytics-api
There is no need for a separate Redis-Query connection as analytics-node
uses the same redis instance for Redis-UVE and Redis-Query. Even if
analytics-api were to report the status of Redis-Query connection, the
status gets updated only on receiving a query from the client, which is
not correct. Since Redis-UVE connection captures the status of the
redis-server, removed the Redis-Query connection from the
process_status.

Change-Id: Ifb917a49642e53412cea4b551da7cc6bd9894449
Closes-Bug: #1560263