Modify flow-aging to ensure upper-bound on time to delete a flow
The flow-aging algorithm depends on two parameters - aging-time and
number of active flows. Flow aging module scans flow-table in a timer.
The flow-aging algorithm tunes the timer-interval and the number of
flows to visit per interval based on flow-table size and flow-age
interval. The current flow-aging algorithm is very conservative in
computing the flow-aging timer and number of flows to visit per
interval. It doesnt guarantee any upper-bound on time to delete an
inactive flow.

The flow aging algorithm is modified as below,
    - Flow aging time is user-defined (default 3mins)
    - A timer is defined to fire every 50 msec
    - The timer scans complete flow-table every scan-interval
      - scan-interval is set as percentage of flow-age time (25% by
                                                             default)

      This ensures that every flow is visited atleast once in
      scan-interval and max-deviation to delete a flow is scan-interval.

      In addition, the flow-aging runs in parallel to both ksync and
      flow-event threads.

Conflicts:
	src/vnsw/agent/vrouter/flow_stats/flow_stats_collector.cc

Change-Id: Ib40680823b649b2d41441aa050d2b4ff75b899d1
Closes-Bug: #1562798