Do not create right service RI for nat instances
If a service instance is in-network-nat mode, the traffic on the right side
is always routed in the primary RI. The service RI is not used for anything.
However, since all routes from primary RIs will still be copied into it.
With this commit, we won't create the right RI for such instances.

Change-Id: I78c16e0d95f5d3799daaf92765da67c51f25fd50
Closes-Bug: 1562200