Added service-health-check-ip property to instance-ip
Currently, vrotuer uses the DNS address for health check packets. That does not
work when there are more than one service instances since each vrouter
advertises the DNS address. To solve this problem, we will reserve a new address
per interface that is connected to a service-health-check object. Agent will use
this ip as source address for the health check packets.

In future, we could optimize by using the service-instance-ip also as
service-health-check-ip if any of them are true:
1. The service instance is in-network-nat
2. There are more than one service instances in the service chain
3. The destination for service-health check is within the data center

Change-Id: I0e89f80059b5711bd6c7fa0ad5da784753d7d18f
Partial-Bug: 1555772