Fix Agent crash in FlowStatsCollector::DeleteEvent
Issue
-----
FlowEntry was pointing to deleted and freed FlowStatsCollector. In this state
when DeleteEvent was received we enqueued Delete Request on stale
FlowStatsCollector which resulted in crash.

Fix
---
When DeleteEvent is received by FlowStatsManager after enqueuing it to
FlowStatsCollector, reset the FlowStatsCollector pointer of FlowEntry. This will
be set when AddEvent is received

Change-Id: Ib7b0d304ff65cde77e535471533b73526ecde339
Closes-Bug: #1577084