Add sequence number support for update queue
Add sequence number support for update queue and show it in introspect. Add
test cases for each API in the update queue class.

The sequence number is 64bits and starts at 1. Each new addition to the end of
the queue gets the next incremented value. If an item is inserted at a non-end
position (which happens only with markers), it gets the sequence number of its
successor, if one exists; else it gets the next incremented value.

Change-Id: I0f9c6a4614d1644e8acf5c05669214d8f63a54e9
Closes-Bug: #1578436