Fix agent crash due to Arithmetic Exception.
Issue
———
Agent crashed at CompositeNH::hash because of Arithmetic exception caused by divide by 0.
In hash API we divide by the size of component NH list of CompositeNH. Crash has happened
because the size of the component NH list is 0.

Fix
——
Add a check to verify the size of component NH list of CompositeNH before using it for dividing.

Change-Id: Ie1d38269935becfdeccd33b144372b202cb59ffd
Closes-Bug: #1536120