Fix interop issue for EVPN w/ MPLS in GRE encapsulation
Current version of EVPN overlay draft says to use MPLSoGRE codepoint,
whereas Contrail uses GRE. This results in interop issues with other
implementations.

The justification for definining a new code point for MPLSoGRE is not
clear. In the intrest of improving interoperability, advertise both GRE
and MPLSoGRE for EVPN routes and map both GRE and MPLSoGRE to the same
string value when advertising routes via xmpp.

Change-Id: I6d01053de225c8058756f9d76682164b6499352e
Closes-Bug: 1588554