Allow the link table introspect to filter entries based on the metadata field
Add another search box where the user enters the substring that he wants the
metadata field to match with.

The input in the search string box will be used as a substring to search the
names of the left and right nodes as before. The input in the new metadata
search box will be used as a substring to match the metadata field.

1. If both the search boxes are filled in, they must both match for the output
to be non-empty.
2. If either (but not both) of the search boxes are non-empty, that particular
search string will be used to filter output entries.
3. If both the search boxes are empty, all the entries will be displayed in the
output.

Change-Id: Idc263110ec9ef4743c886260369700add6c0b18b
Closes-Bug: #1588971