Allow xmpp items with different attributes in same message
UpdateQueue and RibOutUpdates classes are organized to pack BGP NLRI
that share the same attributes into a single message. An unnecessary
side effect is that the same also happens for XMPP.  This results in
more messages, socket write operations and context switches than is
really necessary.

As each XMPP item is self-contained, items with different attributes
can and should be packed into the same XMPP message. This is achieved
by ignoring BgpAttr when comparing UpdateInfos while inserting into a
UpdateQueue and when traversing the UpdateQueue by Attr.

Change-Id: I63a7a7a1b56f31b9e365d27021a60f7a323a197c
Partial-Bug: 1591399