Handle alarm config UPDATE/DELETE in alarm-gen
- Moved the alarm processing code from handle_uve_notif() to a new method
examine_uve_for_alarms() so that it may be invoked on alarm config changes.
- Upon receiving config notification on rabbitmq bus, AlarmGenConfigHandler
would call alarm_config_change_callback() method with a map containing table
name and the alarm that is created/updated/deleted.
Example => If a new alarm 'alarm1' is created with uve_keys
[analytics-node, config-node], then alarm_config_change_callback()
would be called with {'ObjectCollectorInfo': {'alarm1': 'CREATE'},
'ObjectConfigNode': {'alarm1': 'CREATE'}}
- Introduced alarm_config_change_worker() which runs in a separate gevent for
each partition. This gevent would be instantiated in run_uve_processing.
- Handled enable/disable of alarm config
- Presently, examine_uve_for_alarms() is invoked on alarm config change
which executes all the alarms configured for the table. However, for
alarm config changes, only the alarms in the alarm_config_change_map need to
be executed. This optimization will be done in the next commit.

Change-Id: I8ba594ac36b543a21ff06b6b615c60f87c048631
Partial-Bug: #1561596