Fixes: Not sending IdPerms and Perms2 to IFMAP improves perf. API Server Performance Optimization Placeholder bug
Certain common fields like IdPerms and perms2 are not relevant
to be published or published completely to ifmap.

We are still publishing it to IfMAP since they are part of the
schema. Addiing infra to skip publishing fields not relevant
to IfMap completely or partially publish it to IfMap.

For 20k VMs, 20k VMIs, and 60k objects in total, the Init time
stands at 1:28

Change-Id: Ib36aafd4c612472f9d7fc53b74274dbc62a67f7c
Partial-Bug: 1589014