[config] Synchronously allocate security group ID
To improve config node scalability, that patch moves the security group
ID allocation from the schema to the vnc API server.
To prevent any issue during the upgrade from version where sg ID is
allocated by the schema to the version where it's allocated by the vnc
API, that patch keeps allocation into the schema transformer code for
one release overlap. It also reverses the virtual network ID allocation
on notification and re-add it into the schema for the same reason.

Closes-bug: #1602285
Change-Id: I0d7c1d724194aa3b5bd3f57fec5d132978af5699