Resync ifmap graph when restart
1/ When a ifmap server restarting, the api reseed all graph with data
fetched from database without the need to restart vnc api server.
2/ If the connection between the api and ifmap server gone down, the api
buffered messages in the ifmap sending queue and send them when the
connection comes up again. If the buffered queue was exceeded, requests
continue to enqueue in the rabbitmq until ifmap server comes up again.

Also fix an issue in the db-loader script to support new keyspace dict
and add an option to not load zookeeper database.

Change-Id: I8123deed561c9344b854e65b60b5735a2ba4d19e
Closes-Bug: #1622101