Send PollRequest prior to reading PollResponse
Since reading a PollResponse is an expensive operation, send the next
PollRequest to the IFMap server before performing the read. This lets
the IFMap server start building the next PollResponse while we are
processing the previous one.

Note that this means that we do not go back to SendPoll state after
entering PollResponseWait. We stay in PollResponseWait as long we can
successfully send PollRequests and read PollResponses.

This optimization results in a 15% reduction in the time required to
get initial configuration from IFMap server in scaled scenarios.

Change-Id: I5599c897f13ee3a256b2c6b776b4dc69e665af1a
Partial-Bug: 1596089