Change initial state of health check to active
Issue:
------
Health check instance starts with state as inactive,
which cause disruption for traffic when a new health
check service is attached to service instance causing
existing traffic to drop even if the first run of
health check probe return success.

Fix:
----
change the initial state of health check to be active
so that it does not cause disruption in existing traffic
while attaching a new health check service.

Closes-Bug: 1610930
Change-Id: I8f3a9101119af307ba0069fa24499f8d192d9468