Retain unique route attributes when packing into a xmpp message
This bug is a regression introduced by changes made for bug 1591399.

Since an xmpp message now contains routes with different attributes,
Community and ExtCommunity need to processed for each route that is
added to the message.  Otherwise, all routes that get packed into a
message will accidentally be advertised with some of the attributes
of the first route in the message.

Change-Id: Id34e32e3ff106e97c2536b1d0a72772387ca6a76
Closes-Bug: 1612766