Fix alarm processing for virtual-machine object configured under project
Alarms configured under project gets applied to the objects configured
under that project. It is done based on the prefix match of the project
fqname in the uve name (key). For virtual-machine UVE, the name is uuid
and hence it requires special handling in the alarm processing code.
project-name for virtual-machine object is extracted from
UveVirtualMachineAgent.interface_list to compare with the parent fqname
of the alarm configurred under project.

Change-Id: I9fbe5bc8574da5daa87f92e36778753af4186d67
Closes-Bug: #1615091