Redesign update generation code to improve performance
The SchedulingGroupManager design allows multiple SchedulingGroups to
work in parallel. Each SchedulingGroup produces and sends updates for
the RibOuts and Peers that belong to it. A SchedulingGroup is created
and updated such that all for all Peers, all RibOuts they advertise are
part of the same SchedulingGroup and for all RibOuts, all their member
Peers are part of the same SchedulingGroup.

The above design works well when there are many scheduling groups since
each of them can work in parallel.  However, it does not work well for
xmpp as we create a single RibOut per table for xmpp. If there's a large
number of xmpp peers and a large number of tables, we effectively end
up with a single SchedulingGroup for all xmpp peers.  This becomes the
main bottleneck in highly scaled scenarios with a large number of routes
and vrouters.

Implement a new design that creates a BgpSenderPartition per database
partition. A BgpSenderPartition effectively has the same functionality
as a SchedulingGroup. The main difference is that a BgpSenderPartition
contains all RibOuts and Peers - it's not created, updated and deleted
based on Peers joining/leaving RibOuts.

Additionally, shard the RibOutUpdates (and hence the UpdateQueues that
are contained within) for each RibOut instead of using a single set of
UpdateQueues for each RibOut.

The new design allows multiple BgpSenderPartitions to work in parallel
on the corresponding RibOutUpdates shard.

There are some downsides to the new design:

1. Update packing based on common attributes (applicable only for bgp)
is now limited to the scope of a shard instead of being done across the
board.

2. It can result in some lock contention if multiple BgpSenderPartitions
attempt to write to the same peer. This will not happen too often with a
large number of peers.

3. The amount of state compression is reduced to some extent since each
BgpSenderPartition independently discovers that a peer is blocked when
writing to the peer.

The benefits of being able to build and send updates for each shard in
parallel outweigh the downsides. Further, this design also allows (still
to be implemented) the possibility of not using any locks when producing
and consuming updates to/from UpdateQueues by making the producer and
consumer Tasks for any shard mutually exclusive.

This patch implements the following changes:

o Move code from scheduling_group.[cc|h] to bgp_update_sender.[cc|h] and
surgically remove all the SchedulingGroup split and merge related logic.

o The simplified SchedulingGroupManager is now called BgpUpdateSender.

o The simplified SchedulingGroup is now called BgpSenderPartition.

o Allocate N RibOutUpdates per RibOut based on number of DB partitions.

o Modify BgpExport to use the RibOutUpdates for the appropriate shard.

o Make RibOutUpdates::Queue[Join|Leave] happen automatically as part of
RibOut::Regsister|Unregister instead of being called explicitly by the
caller.

o Rename bgp::SendTask to bgp::SendUpdate.

o Remove the tests in scheduling_group_test.cc since they are no longer
required - they covered SchedulingGroup split and merge functionality.

o Move the code in bgp_sg_test.cc to bgp_update_sender_test.cc and tweak
it suitably to test one BgpSenderPartition instead of a SchedulingGroup.

o Modify all update related tests to use BgpSenderPartition at index 0.
Also modify the tests to allocate RibOut dynamically instead of using a
member defined in the unit test class.  This is needed to so that peers
can be cleanly unregistered from RibOuts during test teardown.  This in
turn allows BgpSenderPartition to assert that all state has been cleaned
up properly when it's destroyed.

Change-Id: I848988b07bfd025c62f1c69bf4ddaa38e888e142
Partial-Bug: 1607132