Agent::Diag task policy to be updated
Agent::Diag task is using Find() api for db-tables and is not marked
with mutual-exclusion policy for db-table task. Find() api is using a
scope-lock currently, which is fine. But since it is being evaluated
for removing scope-lock in Find() api, Diag task in Agent should be
marked with mutual-exclusion policy for db-table task. Other Agent tasks
which use Find() api are already marked with this policy.

Change-Id: I436e9f94b7030cbfd3738bd19086ccadecaba379
Partial-Bug: #1611196