Eliminate use of locks in update generation code
Make db::DBTable and bgp::SendUpdate mutually exclusive for a given
partition. Since the RibOutUpdates for a given partition only has a
single producer/consumer (db::DBTable/bgp::SendUpdate respectively),
we can remove the use of locks in UpdateQueue and RibUpdateMonitor.

In theory, we can get rid of RibUpdateMonitor altogether, but we're
leaving it alone for now.

Change-Id: I854ec4b2f6dbe94af4460c1254e5e5d5c9eae187
Partial-Bug: 1607132