Make test_sandesh_kstate tests generic
The UT test_sandesh_kstate sends introspect requests for details of various objects of vrouter. The UT checks the number of responses and the number of objects received. The number of responses varies as and when the size of the object/data-structure is increased/decreased. To keep this UT immune from the changes in the size of data-structures we should only verify the number of objects received. We need to skip the validation of number of messages received.

Change-Id: I219e86cd164cf74cf55bc64248cfc46bbaf917ca
Closes-Bug: #1625922